Publish `py` modules, rearrange `TermId`.
